# web-app/openapi.yaml
openapi: 3.1.0
info:
  title: GEO-SEO Domination API
  version: 1.0.0
  description: |
    RESTful API for SEO analysis, ranking tracking, CRM orchestration, and multi-tenant project management.

    ## Authentication
    All endpoints require a valid JWT bearer token obtained via Supabase Auth.

    ## Multi-Tenancy
    All data is scoped to the user's current organisation. Users must be members of an organisation to access data.

    ## Rate Limits
    - Free: 10 requests/hour
    - Starter: 100 requests/hour
    - Pro: 1,000 requests/hour
    - Enterprise: 10,000 requests/hour

  contact:
    name: API Support
    email: support@geo-seo-domination.com
  license:
    name: Proprietary
    url: https://geo-seo-domination.com/terms
servers:
  - url: https://geo-seo-domination.vercel.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Development
...
paths:
...
  /organisations/switch:
    post:
      summary: Switch organisation context
      tags: [Organisations]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organisationId]
              properties:
                organisationId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Organisation switched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  organisationId:
                    type: string

  # Webhooks
  /webhooks/github:
    post:
      summary: GitHub webhook endpoint
      description: Handles GitHub events (push, pull_request, issues, release)
      tags: [GitHub]
      security: []  # Uses HMAC signature instead of bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitHub webhook payload (varies by event type)
      responses:
        '200':
          description: Webhook processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  deliveryId:
                    type: string
                  eventType:
                    type: string
        '403':
          description: Invalid signature
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
...